Feature/coverage report

+++ gdart.sh
@@ -13,8 +13,7 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-
-OFFSET=$(dirname $BASH_SOURCE[0])
+OFFSET=$(dirname $BASH_SOURCE)
 if [[ -z "$OFFSET" ]]; then
     OFFSET="."
 fi
@@ -24,7 +23,7 @@ source $OFFSET/config
 SPOUT="$GRAALVM_HOME/java"
 JAVAC="$GRAALVM_HOME/javac"
 
-DSE="-Dconcolic.execution=false"
+DSE="-Dconcolic.execution=true" # WICHTIG
 TAINT="-Dtaint.flow=OFF"
 SOLVER="-Ddse.dp=z3"
 SOLVER_FLAGS="-Ddse.witness=false -Ddse.b64encode=true" 
@@ -45,6 +44,7 @@ function usage() {
   echo "    dse.bounds.iter       no. of bounded solving attempts before dropping bounds"
   echo "    dse.bounds.type       fibonacci: uses fibonacci seq. from index 2 (1, 2, 3, 5, ...) as steps"
   echo "    jconstraints.multi=disableUnsatCoreChecking=[true|false]"
+  echo "    dse.coveragereport    generate coverage report: true / false (default)"
   echo ""
 }
 
@@ -56,7 +56,7 @@ while [[ $# -gt 0 ]]; do
       DSE="-Dconcolic.execution=true"
       shift # past argument
       ;;
-    -t|-taint)
+    -t|--taint)
       TAINT="-Dtaint.flow=$2"
       shift # past argument
       shift # past value
@@ -90,7 +90,7 @@ set -- "${POSITIONAL_ARGS[@]}" # restore positional parameters
 
 SOLVER_FLAGS="$SOLVER_FLAGS $SOLVER" 
 
-tmpdir=`mktemp -d`
+tmpdir=$(mktemp -d)
 classpath=$tmpdir
 mainclass=$1
 mainjava=$tmpdir/$1.java
@@ -103,6 +103,12 @@ for cpelement in $@; do
     classpath="$classpath:$cpelement"
   fi
 done
+
+if [[ "$OSTYPE" == "darwin"* ]]; then
+  find $classpath -name "*.java" -exec sed -i "" -e "s/org\.sosy_lab\.sv_benchmarks\.Verifier/tools\.aqua\.concolic\.Verifier/g" {} \;;
+else
+  find $classpath -name "*.java" -exec sed -i "s/org\.sosy_lab\.sv_benchmarks\.Verifier/tools\.aqua\.concolic\.Verifier/g" {} \;;
+fi
 
 classpath="$classpath:$OFFSET/verifier-stub/target/verifier-stub-1.0.jar"
 
@@ -133,10 +139,38 @@ if [[ $? -ne 0 ]]; then
   exit 1
 fi
 
-echo "invoke DSE: $JAVA -cp $OFFSET/dse/target/dse-0.0.1-SNAPSHOT-jar-with-dependencies.jar tools.aqua.dse.DSELauncher $SOLVER_FLAGS -Ddse.executor=$OFFSET/executor.sh -Ddse.executor.args=\"-cp $classpath $DSE $TAINT $mainclass\""
+
+echo "Solver Flags: $SOLVER_FLAGS"
+# echo "invoke DSE: $JAVA -cp $OFFSET/dse/target/dse-0.0.1-SNAPSHOT-jar-with-dependencies.jar tools.aqua.dse.DSELauncher $SOLVER_FLAGS -Ddse.executor=$OFFSET/executor.sh -Ddse.executor.args=\"-cp $classpath $DSE $TAINT $mainclass\""
+echo "invoke DSE $JAVA -cp $OFFSET/dse/target/dse-0.0.1-SNAPSHOT-jar-with-dependencies.jar tools.aqua.dse.DSELauncher $SOLVER_FLAGS -Ddse.executor=$OFFSET/executor.sh -Ddse.executor.args="-cp $classpath $DSE $TAINT $mainclass" -Ddse.sources=$classpath > _gdart.log 2> _gdart.err"
 $JAVA -cp $OFFSET/dse/target/dse-0.0.1-SNAPSHOT-jar-with-dependencies.jar tools.aqua.dse.DSELauncher $SOLVER_FLAGS -Ddse.executor=$OFFSET/executor.sh -Ddse.executor.args="-cp $classpath $DSE $TAINT $mainclass" -Ddse.sources=$classpath > _gdart.log 2> _gdart.err
 
-#Eventually, we print non readable character from the SMT solver to the log.
+# Ensure coverage directory exists
+COVERAGE_DIR="$OFFSET/coverage-report"
+mkdir -p "$COVERAGE_DIR"
+
+# Copy original and test classes to the new Maven project
+RELEVANT_CLASSES_DIR="$COVERAGE_DIR/src/main/java"
+mkdir -p "$RELEVANT_CLASSES_DIR"
+cp $mainjava "$RELEVANT_CLASSES_DIR/"
+
+# Maven build and test in the new project
+POM_PATH="$COVERAGE_DIR/pom.xml"
+if [[ ! -f $POM_PATH ]]; then
+  echo "POM file $POM_PATH does not exist. Exiting..."
+  exit 1
+fi
+
+echo "Running Maven build and tests in the new project"
+mvn clean test -f $POM_PATH
+
+# Check if coverage report is enabled
+if [[ "$SOLVER_FLAGS" == *"dse.coveragereport=true"* ]]; then
+  echo "Generating JaCoCo coverage report"
+  mvn jacoco:report -f $POM_PATH
+fi
+
+# Eventually, we print non-readable characters from the SMT solver to the log.
 sed 's/[^[:print:]]//' _gdart.log > _gdart.processed
 mv _gdart.processed _gdart.log
 
@@ -150,18 +184,18 @@ cat _gdart.err
 
 echo "# # # # # # #"
 
-complete=`cat _gdart.log | grep -a "END OF OUTPUT"`
+complete=$(cat _gdart.log | grep -a "END OF OUTPUT")
 if [[ $complete = "" ]]; then 
   complete="no" 
 else 
   complete="yes" 
 fi
 
-errors=`cat _gdart.log | grep -a ERROR | grep -a java.lang.AssertionError | cut -d '.' -f 3 | wc -l | tr -s '[:blank:]'`
-buggy=`cat _gdart.log | grep -a BUGGY | cut -d '.' -f 2 | wc -l | tr -s '[:blank:]'`
-diverged=`cat _gdart.log | grep -a DIVERGED | cut -d '.' -f 2 | wc -l | tr -s '[:blank:]'`
-skipped=`cat _gdart.log | grep -a SKIPPED | egrep -v "assumption violation" | cut -d '.' -f 3 | wc -l | tr -s '[:blank:]'`
-taint=`cat _gdart.log | grep -a "TAINT VIOLATION" | egrep -v "assumption violation" | cut -d '.' -f 3 | wc -l | tr -s '[:blank:]'`
+errors=$(cat _gdart.log | grep -a ERROR | grep -a java.lang.AssertionError | cut -d '.' -f 3 | wc -l | tr -s '[:blank:]')
+buggy=$(cat _gdart.log | grep -a BUGGY | cut -d '.' -f 2 | wc -l | tr -s '[:blank:]')
+diverged=$(cat _gdart.log | grep -a DIVERGED | cut -d '.' -f 2 | wc -l | tr -s '[:blank:]')
+skipped=$(cat _gdart.log | grep -a SKIPPED | egrep -v "assumption violation" | cut -d '.' -f 3 | wc -l | tr -s '[:blank:]')
+taint=$(cat _gdart.log | grep -a "TAINT VIOLATION" | egrep -v "assumption violation" | cut -d '.' -f 3 | wc -l | tr -s '[:blank:]')
 
 echo "analysis completed: $complete"
 
@@ -170,6 +204,8 @@ printf 'taint discovered (data/control):  %4s\n' $taint
 printf 'crashes in concolic executor:     %4s\n' $buggy
 printf 'unexpected paths for models:      %4s\n' $diverged
 printf 'skipped (assumptions/unsupported):%4s\n' $skipped
+
+echo "Coverage report generated at $COVERAGE_DIR/target/site/jacoco"
 
 rm -rf $tmpdir 
 rm _gdart.log
